Refactor license update logic, extract fetch logic to a file
