+++ lib/create-licenses.js
@@ -109,7 +109,7 @@ const developLicenses = pkgs => {
 
     if (licenseText) {
       developed.push({
-        ...pkg,
+        name: pkg.name,
         licenses,
         licenseText: licenseText,
       });
